requirements.txt
discord.py
flask
matplotlib
